switch notification streaming: SendFrame -> SendJSON
