+++ package.json
@@ -13,10 +13,10 @@
   "author": "Niclas Hedhman",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@grafana/data": "8.4.4",
-    "@grafana/runtime": "8.4.4",
-    "@grafana/toolkit": "8.4.4",
-    "@grafana/ui": "8.4.4",
+    "@grafana/data": "8.5.0",
+    "@grafana/runtime": "8.5.0",
+    "@grafana/toolkit": "8.5.0",
+    "@grafana/ui": "8.5.0",
     "@types/lodash": "^4.14.172"
   },
   "resolutions": {
